fixed issue of setAccessToken called twice
