+++ frontend/src/components/Home/Home.js
@@ -1,4 +1,6 @@
-import React from "react";
+import React, { useEffect } from "react";
+
+import { useAuth0 } from "../../react-auth0-spa";
 
 import "./Home.css";
 
@@ -10,6 +12,26 @@ const gridRedirects = ["/about", "/projects", "/blog"];
 let gridItems;
 
 function Home() {
+  const { isAuthenticated, user } = useAuth0();
+
+  // should be fetching on load of the Home page and NOT the Navigation bar (Website component)
+  // that was my mistake earlier
+  useEffect(() => {
+    if (isAuthenticated && user) {
+      const body = {
+        uid: user.sub
+      };
+      fetch("http://localhost:3001/setAccessToken", {
+        method: "POST",
+        headers: {
+          Accept: "text/html",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify(body)
+      });
+    }
+  });
+
   gridItems = [];
   for (let [i, lbl] of gridLabels.entries()) {
     let cssComponent = classNames("homeGrid", "homePageLinks", gridCSS[i]);
